prepare return

+++ interfaces/inverse.sol
@@ -197,7 +197,7 @@ interface CErc20Interface is CTokenInterface {
     /*** User Interface ***/
     function mint(uint mintAmount) external returns (uint);
     function redeem(uint redeemTokens) external returns (uint);
-    function redeemUnderlying(uint redeemAmount) external returns (uint);
+    function redeemUnderlying(uint redeemAmount) external returns (uint256);
     function borrow(uint borrowAmount) external returns (uint);
     function repayBorrow(uint repayAmount) external returns (uint);
     function repayBorrowBehalf(address borrower, uint repayAmount) external returns (uint);
